+++ Editor/ActionWindow/Cleanup/CleanupExtensions.cs
@@ -165,5 +165,66 @@ namespace Vertx.Editor
 		}
 
 		#endregion
+
+		#region Scripts
+
+		private const string cleanupMissingComponentsTitle = "Component - Remove Missing Components";
+
+		[ActionProvider]
+		private static ActionOperation CleanupMissingScriptsComponentsInProject()
+		{
+			return new ActionOperation(ActionOperation.ActionTarget.All,
+				cleanupMissingComponentsTitle,
+				"Component", () =>
+				{
+					using (var scope = new EditorUtils.BuildSceneScope())
+					{
+						foreach (var scene in scope)
+						{
+							IEnumerable<GameObject> gameObjects = EditorUtils.GetAllGameObjectsInScene(scene);
+							int countRemoved = RemoveMissingScriptsComponentsOnGameObjects(gameObjects);
+							if (countRemoved > 0)
+							{
+								Debug.Log($"Removed {countRemoved} GameObject names in {scene.name}");
+								EditorSceneManager.SaveScene(scene);
+							}
+						}
+					}
+				});
+		}
+
+		static int RemoveMissingScriptsComponentsOnGameObjects(IEnumerable<GameObject> gameObject)
+		{
+			int countRemoved = 0;
+			foreach (GameObject gO in gameObject)
+			{
+				var components = gO.GetComponents<Component>();
+				foreach (Component component in components)
+				{
+					if (component == null)
+					{
+						SerializedObject sO = new SerializedObject(gO);
+						var componentsProp = sO.FindProperty("m_Component");
+						for (int i = components.Length - 1; i >= 0; i--)
+						{
+							//If it's a prefab and connected to it, then we should be modifying the prefab instead.
+							if(PrefabUtility.GetPrefabInstanceStatus(components[i]) != PrefabInstanceStatus.NotAPrefab) continue;
+							if (components[i] == null)
+							{
+								componentsProp.DeleteArrayElementAtIndex(i);
+								countRemoved++;
+							}
+						}
+
+						sO.ApplyModifiedPropertiesWithoutUndo();
+						break;
+					}
+				}
+			}
+
+			return countRemoved;
+		}
+
+		#endregion
 	}
 }