+++ Editor/Core/EditorUtils.cs
@@ -7,6 +7,7 @@ using System.Reflection;
 using System.Text;
 using System.Text.RegularExpressions;
 using UnityEditor;
+using UnityEditor.SceneManagement;
 using UnityEngine;
 using UnityEngine.SceneManagement;
 using Object = UnityEngine.Object;
@@ -342,6 +343,56 @@ namespace Vertx.Extensions
 
 		#region Scene
 
+		public class BuildSceneScope : IEnumerator<Scene>, IEnumerable<Scene>
+		{
+			private readonly SceneSetup[] sceneManagerSetup;
+
+			public BuildSceneScope()
+			{
+				sceneManagerSetup = EditorSceneManager.GetSceneManagerSetup();
+				buildSceneCount = SceneManager.sceneCountInBuildSettings;
+				currentScene = default;
+				progressIncrement = 1 / (float)buildSceneCount;
+			}
+
+			public void Dispose()
+			{
+				EditorUtility.ClearProgressBar();
+				if (sceneManagerSetup != null && sceneManagerSetup.Length > 0)
+					EditorSceneManager.RestoreSceneManagerSetup(sceneManagerSetup);
+			}
+
+			private readonly float progressIncrement;
+
+			public void DisplayProgressBar(string title, string info, float localProgress) 
+				=> EditorUtility.DisplayProgressBar(title, info, (buildIndex + localProgress * progressIncrement) / buildSceneCount);
+
+			private readonly int buildSceneCount;
+			private int buildIndex = -1;
+			private Scene currentScene;
+
+			public bool MoveNext()
+			{
+				//Avoids going beyond the end of the collection.
+				if (++buildIndex >= buildSceneCount)
+					return false;
+
+				string path = SceneUtility.GetScenePathByBuildIndex(buildIndex);
+				EditorSceneManager.OpenScene(path, buildIndex == 0 ? OpenSceneMode.Single : OpenSceneMode.Additive);
+				currentScene = SceneManager.GetSceneByBuildIndex(buildIndex);
+				return true;
+			}
+
+			public void Reset() => buildIndex = -1;
+
+			public Scene Current => currentScene;
+
+			object IEnumerator.Current => Current;
+			public IEnumerator<Scene> GetEnumerator() => this;
+
+			IEnumerator IEnumerable.GetEnumerator() => GetEnumerator();
+		}
+
 		public static IEnumerable<T> GetAllComponentsInScene<T>(Scene scene) where T : Component
 		{
 			GameObject[] roots = scene.GetRootGameObjects();
@@ -357,7 +408,10 @@ namespace Vertx.Extensions
 				//Get components
 				T[] components = @base.GetComponents<T>();
 				foreach (T component in components)
-					yield return component;
+				{
+					if(component != null)
+						yield return component;
+				}
 
 				foreach (Transform child in @base)
 				{
@@ -387,6 +441,77 @@ namespace Vertx.Extensions
 						yield return component;
 				}
 			}
+		}
+
+		#endregion
+
+		#region Logging Extensions
+
+		/// <summary>
+		/// Returns an appropriate full path to the object. This includes the scene if relevant.
+		/// </summary>
+		/// <param name="object">The Object to get a path to</param>
+		/// <returns>Path to the Object</returns>
+		public static string GetPathForObject(Object @object)
+		{
+			bool persistent = EditorUtility.IsPersistent(@object);
+			string path = persistent ? $"{AssetDatabase.GetAssetPath(@object)}/" : string.Empty;
+			switch (@object)
+			{
+				case Component component:
+					{
+						//The component already includes the base child in its ToString function, so we can use the parent.
+						Transform transform = component.transform.parent;
+						string tPath;
+						if (transform != null)
+						{
+							tPath = AnimationUtility.CalculateTransformPath(transform, null);
+							if (persistent)
+							{
+								//For prefabs, the path already includes the root. So we can remove it from the transform path.
+								int indexOf = tPath.IndexOf('/');
+								tPath = indexOf < 0 ? null : tPath.Substring(indexOf + 1);
+							}
+						}
+						else
+							tPath = null;
+
+						var scene = component.gameObject.scene;
+						if (scene.IsValid())
+							path += $"({scene.path}) ";
+						path += string.IsNullOrEmpty(tPath) ? component.ToString() : $"{tPath}/{component}";
+					}
+					break;
+				case GameObject gameObject:
+					{
+						//The gameObject already includes the base child in its ToString function, so we can use the parent.
+						Transform transform = gameObject.transform.parent;
+						string tPath;
+						if (transform != null)
+						{
+							tPath = AnimationUtility.CalculateTransformPath(transform, null);
+							if (persistent)
+							{
+								//For prefabs, the path already includes the root. So we can remove it from the transform path.
+								int indexOf = tPath.IndexOf('/');
+								tPath = indexOf < 0 ? null : tPath.Substring(indexOf + 1);
+							}
+						}
+						else
+							tPath = null;
+
+						var scene = gameObject.scene;
+						if (scene.IsValid())
+							path += $"({scene.path}) ";
+						path += string.IsNullOrEmpty(tPath) ? gameObject.ToString() : $"{tPath}/{gameObject}";
+					}
+					break;
+				default:
+					path += @object.ToString();
+					break;
+			}
+
+			return path;
 		}
 
 		#endregion